refactor(cli): extract unity compose handler
